drm/rockchip: vop: correct RK3288/RK3399 VOP WIN register fields
In RK3288/RK3399, there is a problem with the definition of the
`fmt_10` field in the VOP WIN register. The original `fmt_10`
definition conflicts with the `lb_mode` definition.

This will cause the `lb_mode` function to be affected, there may be
problems with incorrect colors displayed, incorrect window sizes,
image flickering, etc.

This patch correct `fmt_10` field to the expected value in TRM.

Change-Id: I0f9f23c459dc5870532e28b74053a9b4dc606de0
Signed-off-by: Chaoyi Chen <chaoyi.chen@rock-chips.com>